Rename the [I18n]FormatterGlbChecker tests to [I18n]FormatterLubGlbChecker
